requirements.txt
# image conversion libraries
pillow
pillow_heif
